cmd: move config headers to include directory
- Move config headers to include directory and general clean up
  and address review comments
