+++ cmd/img_mgmt/include/img_mgmt/img_mgmt_config.h
